+++ NEWS.md
@@ -1,6 +1,7 @@
 # nswgeo (development version)
 
-* Add Public Health Network (PHN) map data as `phn` dataset.
+* Add Primary Health Network (PHN) map data as `phn` dataset.
+* Fixed the `poa_nsw` dataset to avoid strange geometries near the state border.
 
 # nswgeo 0.3.3